Fixed errors.

+++ src/bin/package/clean.rs
@@ -6,14 +6,16 @@ use std::fs::remove_dir_all;
 
 
 pub(crate) fn clean() -> Result<(), Error> {
-    if let Err(error) = remove_dir_all(DIR_PROJECT_PACKAGES.as_path()) {
-        return Err(Error {
-            message: format!(
-                "failed to remove directory: {}",
-                DIR_PROJECT_PACKAGES.to_string_lossy(),
-            ),
-            source: Some(Box::new(error)),
-        });
+    if DIR_PROJECT_PACKAGES.exists() {
+        if let Err(error) = remove_dir_all(DIR_PROJECT_PACKAGES.as_path()) {
+            return Err(Error {
+                message: format!(
+                    "failed to remove directory: {}",
+                    DIR_PROJECT_PACKAGES.to_string_lossy(),
+                ),
+                source: Some(Box::new(error)),
+            });
+        }
     }
 
     Ok(())
